backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
numpy==1.23.5
pandas==1.5.3
faiss-cpu==1.7.3
sentence-transformers==2.2.2
python-dotenv==1.0.0
canvasapi==3.1.0